[chore] rename image-uploader.ts to useImageUpload.ts
